+++ src/layouters/threeDiffToBaseLayouter.ts
@@ -1,16 +1,17 @@
 import * as vscode from 'vscode';
-import { Monitor } from './monitor';
+import { Monitor } from '../monitor';
 import { DiffLayouter, DiffLayouterFactory } from './diffLayouter';
-import { defaultTemporarySideBySideSettingsManager } from './temporarySettingsManager';
+import { defaultTemporarySideBySideSettingsManagerLazy } from '../temporarySettingsManager';
 import { SplitDiffLayouter, GroupOrientation, diffEditorSymbol } from './splitDiffLayouter';
-import { DiffedURIs } from './diffedURIs';
+import { DiffedURIs } from '../diffedURIs';
 
 export class ThreeDiffToBaseLayouterFactory implements DiffLayouterFactory {
   public readonly settingValue = "3DiffToBase";
 
   public create(
     monitor: Monitor,
-    temporarySideBySideSettingsManager = defaultTemporarySideBySideSettingsManager,
+    temporarySideBySideSettingsManager =
+      defaultTemporarySideBySideSettingsManagerLazy.value,
     diffedURIs: DiffedURIs,
   ): DiffLayouter {
     return new SplitDiffLayouter(monitor, diffedURIs, (diffedURIs) => ({
@@ -42,8 +43,6 @@ export class ThreeDiffToBaseLayouterFactory implements DiffLayouterFactory {
           size: 0.3,
         },
       ]
-    }),
-      temporarySideBySideSettingsManager,
-    );
+    }), temporarySideBySideSettingsManager);
   }
 }