+++ src/lazy.ts
@@ -1,0 +1,16 @@
+export class Lazy<T> {
+  public get value(): T {
+    if (!this.initialized) {
+      this._value = this.factory();
+      this.initialized = true;
+    }
+    return this._value!;
+  }
+
+  public constructor(
+    private readonly factory: () => T,
+  ) { }
+
+  private initialized = false;
+  private _value: T | undefined;
+}