+++ tools/post-commit.ts
@@ -9,7 +9,7 @@ runAsync(async () => {
 
     // Based on: https://stackoverflow.com/a/19328859
     if (await runCommand(git, ["cherry-pick", "-n", "-m1", "-Xtheirs", "stash"]) !== 0) { return 1; }
-    
+
     fs.unlinkSync(".git/MERGE_MSG");
     if (await runCommand(git, ["restore", "--staged", "."]) !== 0) { return 1; }
     if (await runCommand(git, ["stash", "drop", "stash@{0}"]) !== 0) { return 1; }